changed math to use doubles instead of bigdecimel

+++ src/main/kotlin/math/Operator.kt
@@ -1,5 +1,3 @@
 package math
 
-import java.math.BigDecimal
-
-data class Operator(val symbol: Char, val precedence: Int, val operation: Function2<BigDecimal>)
+data class Operator(val symbol: Char, val precedence: Int, val operation: Function2<Double>)
